Updated to the newer JEI
Updated the JEI support to the newer JEI, which changed the API.
Closed #42 as blank recipes don't seem to be created in older JEI
versions with this fix, and tweaked a few things.
However, the bug with craftweaker recipes not being in JEI is still not
fixed, as that seems to be on JEI's side.