Removed (Back) counter gears, bugfixes
Removed (Back) counter gears entirely because there is the issue of them
being imprecise that would require a complete re-write of the rotational
system to fix, namely they rotate too little before stopping. This is
due to partial rotations not being possible, so an entire tick of
spinning has to be cut, even when most of it would be justified.
Considering there are other ways of fufilling the same purpose (spinning
a constant angle before stopping), and these devices could not be
reasonably fixed, they had to be removed.