Re-Wrote magic system internals
Switched to a 1 stage cycle, simplified the code, cleaned it up, and
centrilized it to BeamRenderTE.

Did a lot of tweaking to the prototype loading/unloading code, it still
isn't quite where it needs to be yet though.

There are still some issues with loading/unloading of beams through
prototypes though (mostly about delays being added or beams merging).
This will take a lot of fine tuning to fix.