Stamp Mill
The Stamp Mill is basically functional at this point, only missing an item model, and maybe a sound
The Stamp Mill grinds ores into 3 ore gravels. Unlike most other machine, the Stamp Mill has two progress bars: One based on rotary energy used and the other based on time. The machine does an operation when either bar fills, but if the time bar fills before the progress bar the item is not ground into gravel, and the ore is instead sent to the output unprocessed. Unless (unpractically) high gear speeds are used, it will take multiple runs through the Stamp Mill to process an ore, requiring filtering the output and sending some of the products back into the Mill.