Removed Magic
Admit it, for a moment there you believed this commit's title.

The magic system has been renamed to the beam system. (This has resulted in a large refactor)
Crossroads is no longer a Tech and Magic mod, it is now only a Tech mod. To be clear, the magic system is still in the game, but most everything in it has been renamed to be non-magical.
Too many people got confused by the magical names yet non-magical visual themes/mechanics.

Arcane Extractor->Resonance Extractor
Arcane Reflector->Beam Reflector
Lens Holder->Lens Frame
Magic Elements->Beam Alignments
and so on.

-Also, the Beam Splitter has been split (no pun intended) into two separate blocks: The Beam Siphon and the Beam Redirector. To be clear, the Beam Splitter itself has been removed.

+The Beam Redirector is cheap, and is controlled with a redstone signal to choose between one of two outputs. It can't split a beam into parts at all, but it acts as a simple junction or intersection.

+The Beam Siphon is far more interesting (and comprable in price to the old Beam Splitter). It has two outputs, and the power out the main output is equal to the redstone signal received. If it has a signal of 4, a power four beam comes out the front, and all other magic received comes out the back. It can also use larger values than 15 (via ratiators). It should make building central beam powerlines and massive beam networks far more rewarding and intuitive.

+The Small and Large Quartz Stabilizer have been merged into one block (the Quartz Stabilizer), which has its output configured with a wrench between several values.

*The Quartz Stabilizer now outputs exactly the power set, instead of being an approximation due to rounding errors.