Conduits
Bit the bullet and ported the conduit type blocks (heat cables, pipes, alchemy conduits). Everything in this branch is experimental, but the conduit stuff is doubly so.
Fluid pipes had a rework to only allow bi-directional connection with each other and tanks.
Wrenching fluid pipes should configure both sides of the connection now
All the conduit stuff needs to be heavily tested