Hydroponics Trough
+Added the Hydroponics Trough. It uses either Nutrient Solution or Fertilizer Solution (fertilizer is cheaper) to grow plants very quickly. It can grow plants outside of normal crops, including things like sugar cane, sea pickles, lilypads, and mushrooms. It is fairly fast (config). Most crops (other than mushrooms and the sea crops) require light. Bonemeal or a potential beam CAN be used on the trough.
The hydroponics trough CAN NOT have the produced items removed UNLESS it has a redstone signal. However, while it has a redstone signal, all stored fluid is vented and the machine shuts down. As the internal tank is fairly large, and the liquid is quite valuable (~40 bonemeal worth fits in the buffer), automating it in a way to minimize waste is valuable.
*Made Nutrient Solution and Fertilizer Solution more expensive
*Tweaked fluid tanks so that the NBT-tag option for creative item picking (it's a vanilla feature people don't realize exists) works to store their contents.